pyproject.toml
[tool.poetry]
name = "hippius"
version = "0.2.70"
description = "Python SDK and CLI for Hippius storage"
authors = ["Dubs <dubs@dubs.rs>"]
readme = "README.md"
repository = "https://github.com/thenervelab/hippius-sdk"
documentation = "https://github.com/thenervelab/hippius-sdk/docs"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]
packages = [{include = "hippius_sdk"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
substrate-interface = "==1.7.11"
httpx = "^0.28.1"
pydantic = "^2.0.0"
python-dotenv = "^1.0.0"
base58 = "^2.1.1"
pynacl = "^1.5.0"
cryptography = ">=46.0.5"
pyperclip = {version = "^1.8.2", optional = true}
mnemonic = "^0.20"
rich = "^14.0.0"
click = "^8.1"

[tool.poetry.extras]
clipboard = ["pyperclip"]

[tool.poetry.scripts]
hippius = "hippius_sdk.cli:main"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
pytest-asyncio = "^0.21.0"
pytest-timeout = "^2.4.0"
black = "^23.0.0"
isort = "^5.12.0"
mypy = "^1.0.0"
pylint = "^2.17.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
